feat(): add rules file and comments
