+++ packages/common/src/core/schemas/internal/metrics/interfaces.ts
@@ -1,0 +1,27 @@
+import { UiSchemaRuleEffects } from "../../types";
+
+export enum LAYOUTS {
+  simple = "simple",
+  dataViz = "dataViz",
+  moreInfo = "moreInfo",
+}
+export enum SCALE {
+  small = "s",
+  medium = "m",
+  large = "l",
+}
+export enum UNIT_POSITIONS {
+  before = "before",
+  after = "after",
+  below = "below",
+}
+export enum VISUAL_INTEREST {
+  none = "none",
+  sparkline = "sparkline",
+  icon = "icon",
+  logo = "logo",
+}
+export enum SOURCE {
+  dynamic = "dynamic",
+  static = "static",
+}
