+++ packages/common/src/core/schemas/internal/metrics/rules.ts
@@ -1,32 +1,10 @@
 import { UiSchemaRuleEffects } from "../../types";
 
-export enum LAYOUTS {
-  simple = "simple",
-  dataViz = "dataViz",
-  moreInfo = "moreInfo",
-}
-export enum SCALE {
-  small = "s",
-  medium = "m",
-  large = "l",
-}
-export enum UNIT_POSITIONS {
-  before = "before",
-  after = "after",
-  below = "below",
-}
-export enum VISUAL_INTEREST {
-  none = "none",
-  sparkline = "sparkline",
-  icon = "icon",
-  logo = "logo",
-}
-export enum SOURCE {
-  dynamic = "dynamic",
-  static = "static",
-}
+/**
+ * Shared metric rules for uiSchema
+ */
 
-/***************** Rules *****************/
+/** Show if the type is set to static */
 export const SHOW_FOR_STATIC_RULE_ENTITY = {
   condition: {
     scope: "/properties/_metric/properties/type",
@@ -35,6 +13,7 @@ export const SHOW_FOR_STATIC_RULE_ENTITY = {
   effect: UiSchemaRuleEffects.SHOW,
 };
 
+/** Show if the type is set to dynamic */
 export const SHOW_FOR_DYNAMIC_RULE_ENTITY = {
   condition: {
     scope: "/properties/_metric/properties/type",
@@ -43,6 +22,7 @@ export const SHOW_FOR_DYNAMIC_RULE_ENTITY = {
   effect: UiSchemaRuleEffects.SHOW,
 };
 
+/** Show if sharing is enabled */
 export const SHOW_FOR_SHARING_RULE_ENTITY = {
   condition: {
     scope: "/properties/_metric/properties/shareable",
